Implemented follow/unfollow functionality in Album View Controller
Button text reads Follow if user isn't following yet; and reads Following if they already are upon View did load. If the user  opts to change the value of follow/unfollow by pressing the button then the button will change accordingly and the number of followers in the description label count will increase or decrease by 1.